Re-enable full data access in schema
In particular for `computeAttributeReference`

Also, enable `computeAttributeReference` to return an array, which is
necessary eg when it needs to check a value other than the one for the
key.  This is useful if the property name differs from the key in the
payload (eg annotating keys with a sigil to indicate that they're
references).

This suggests some cleanup (see #76).