Enable more computeAttributeReference arrays
[breaking]

In particular, attribute array references (ie in which
`isAttributeArrayReference` returns `true`), which return `RecordArray`s
and have different set semantics, now work properly when
`computeAttributeReference` returns an array of references.

This is a breaking change as schemas with attribute array references
will now have `computeAttributeReference` called once and must return an
array of references, instead of being called for each element of the
array.

I'd like to completely remove the need for `isAttributeArrayReference`
(see #76).  Perhaps it would be feasible to always return a
`RecordArray` when `computeAttributeReference` returns an array.