Fix regressions &c. in ember 3.1
Simple assignments of CPs to prototypes is no longer supported unless
those objects go through the `mergeMixins` codepath.  This means it's
fine for `EmberObject.extend` but not for es classes.

Similarly, simple assignments to instances of cps may have issues.  It
actually works in 3.1.0-beta.2, but it's not clear whether or not it
should.  This feature may get deprecated and if so it's likely the
deprecation path will be to use `defineProperty`.

See https://github.com/emberjs/ember.js/pull/16270