Do not resolve values from `setUnknownProperty`
Values that are set on the model directly (as opposed to ones that come from the
server payload) are treated as resolved.  This means that they do not go
through hooks like `schema.computeAttributeReference` for arrays of
references.

This also fixes a bug where setting an already-resolved `RecordArray`
would fail to update the `RecordArray` in-place if the new value was a
`RecordArray` of models (rather than a plain array of models).