Associate resolved values with tracked arrays
In general this is not recommended, but it can help if it's difficult to
determine whether the server payload should be interpreted as a tracked
array or a record array (ie whether it's an array of nested models or a
reference array).  If this is a problem it is most likely a problem for
empty arrays.

Associating resolved values with tracked arrays means that in the
tracked array case new records that are added will be removed when
they're unloaded.