Schema as a service
Instead of registering the schema via a global it is now looked up at
`service:m3-schema` and can therefore participate in the rest of the
container infrastructure.

Among other things, this makes it easier to manage state in schemas, as
well as have app schemas depend on other services.