Fixes rollbackAttributes for record data from projections
Issue:
Calling rollbackAttributes on a record data that has a base record data
was not working properly because the attributes being rolledback were in
the record data from the projection and not the base record data.

When dealing with projections all the interactions should be delegated
to the base record data.

In this PR I am adding a fix to call rollback attributes on the base
record data when one is present.