Test rollbackAttributes on MegamorphicModel
I added another test to verify the behavior of rollbackAttributes on a
MegamorphicModel. This test resembles more the actual behavior in an
application instead of just asserting that a method was called.

This test showcased another issue too in `isAttrDirty` which was
also not calling its base record data to check if an attribute was
dirty.