Fixes rollbackAttributes for record data from projections (#223)
* Fixes rollbackAttributes for record data from projections

Issue:
Calling rollbackAttributes on a record data that has a base record data
was not working properly because the attributes being rolledback were in
the record data from the projection and not the base record data.

When dealing with projections all the interactions should be delegated
to the base record data.

In this PR I am adding a fix to call rollback attributes on the base
record data when one is present.

* Test rollbackAttributes on MegamorphicModel

I added another test to verify the behavior of rollbackAttributes on a
MegamorphicModel. This test resembles more the actual behavior in an
application instead of just asserting that a method was called.

This test showcased another issue too in `isAttrDirty` which was
also not calling its base record data to check if an attribute was
dirty.

* Fix test title

* Adds test for isDirty on a projection

* Add unit test for record data isAttrDirty

The previous implementation was not delegating to its base record when
dealing with projections. This test ensures that we do not regress on
that behavior.