fix: batch change notifications
Because m3 models are proxies (in that their properties are retrieved
via `unknownProperty`), their properties are treated as volatile.

This means that when chain watchers update, the parent value will be
computed (instead of its cached value used).  In practice, this can
result in reference arrays trying to resolve their references durin ga
`pushPayload` when only part of the payload has been pushed into the
store.

This commit fixes the issue by batching change notifications in `_push`.