feat: batch array changes
Previously only property changes were batched; now array changes in
record arrays are also batched.  Note that this is not the case for
tracked arrays, but this feature should be added there as well.

As part of this change, remove `ArrayProxy` as the base class of
`RecordArray`.  Although our arrays are array proxies, the base class is
not adding value and it's easier to implement all of the logic
ourselves.

It is also a breaking change as we will no longer fire `willChange`
events on record arrays.  It's already not a recommended pattern to rely
on any kind of before observer, and it would cost us more to maintain
enough context to fire these asynchronously.