fix: only store base types in global cache
Previously we would update the global cache each time we saw any type,
causing base & projection types to override each other in the order they
appeared.  This only affected the global cache so its only effect was
when resolving references.  In practice this didn't affect projection
types (since there you tend to resolve references by type and not rely
on the global cache) but it meant that referring to base types when any
projections were in the cache was impractical as you would easily get
projection types from references in base types.

[fix #122]