feat: add debug-adapter for Ember Inspector support
Previously Ember Inspector was defaulting to Ember Data's debug-adapter
to handle m3 model types. However, due to differences in how model types
and attributes are accessed and handled, this did not work.
This implements the baseline functionality of being able to inspect m3
model types and see the attributes and attribute values for each record.