fix: fix hook for `internalModel` removal
This is used to clean up our secondary cache.

Note that the new (and prior) monkeymatching is undesired as it's using
private API in both cases, and `_internalModelsFor` is slated to be
removed as private API (not intimate).

The public API that enables this kind of secondary cache should be used

when we switch to the custom model class API (see #294).

[fix #342]

Co-authored-by: Robert Jackson <me@rwjblue.com>