Remove work around for Ember < 2.12 relying on `setOwner` enumerability
Future versions of Ember are likely to move to true `Symbol` for the
implementation of `setOwner`, and if that does happen this code will no
longer work (because symbols are non-enumerable).

Since this was only required for supporting versions of Ember _older_
than 2.12 (and we don't support those versions any longer) we can
remove.