Adds support for native Proxy behind a canary feature flag
Adds support for wrapping M3 models and arrays in a native Proxy,
allowing them to be accessed using Plain JS (tm) syntax. This is still
a work in progress as some functionality, such as `Array.isArray`, does
not work on the arrays, and additionally it is double wrapping
everything with Ember's own proxy system so it is very expensive
overall to enable perf-wise. As such, it is kept behind a feature flag
that can only be enabled in development mode for testing and
experimentation. Enabling this feature flag will show a warning in the
console to users.