Migrate away from ember-qunit-assert-helpers
ember-qunit-assert-helpers hasn't been updated to work with
ember-qunit@5 (and the usage of QUnit from ember-auto-import), in order
to help users migrate away from it (so that they can test against Ember
4.0+, and test against Embroider) `@ember/test-helpers` started exposing
lower level deprecation/warning APIs that we can leverage. There will
eventually be nicer higher level APIs (like the
assert.expectDeprecation style that we had before) in a future
ember-qunit release, but we can leverage the lower level APIs directly
for now.

Co-authored-by: David J. Hamilton <davidjh@hjdivad.com>