Avoid using window.QUnit directly
We need to setup these properties **before** our assets load (which is
why they are setup in `tests/index.html`) but we can't rely on
`window.QUnit` to be available (you have to `import QUnit from 'qunit'`
to get QUnit now that we are leveraging ember-auto-import).

This migrates to using `location.search` directly, which serves the same
purpose.