Fix Ember.A clobbering
Ember.A is very proxy-unfriendly with Ember classic classes.

This is because MutableArray & friends will setEmberArray during create
-> init.  This will happen before an instance is created, so there's no
straightforward way to call create or init with a proxy as the receiver.

The issue is Ember.A's isEmberArray/setEmberArray implementation works
by checking and setting membership in a WeakMap, so during object
instantiation this will be done by the target and in all other cases it
will be done by the proxy.

All of this makes Ember.A try to apply NativeArray *again*. The reason
this is bad is that NativeArray.apply will indiscriminately clobber
existing methods like objectAt by just assuming the argument is a native
array and nothing that presents as array-like (which may have its own
implementation of functions like objectAt)